+++ src/hdlc_send.go
@@ -60,7 +60,7 @@ func layer2_send_frame(channel C.int, pp *packet_t, bad_fcs C.int, audio_config_
 	} else if audio_config_p.achan[channel].layer2_xmit == LAYER2_FX25 {
 		var fbuf [AX25_MAX_PACKET_LEN + 2]C.uchar
 		var flen = ax25_pack(pp, &fbuf[0])
-		var n = fx25_send_frame(channel, &fbuf[0], flen, C.int(audio_config_p.achan[channel].fx25_strength), false)
+		var n = fx25_send_frame(channel, &fbuf[0], C.int(flen), C.int(audio_config_p.achan[channel].fx25_strength), false)
 		if n > 0 {
 			return (n)
 		}
@@ -72,7 +72,7 @@ func layer2_send_frame(channel C.int, pp *packet_t, bad_fcs C.int, audio_config_
 
 	var fbuf [AX25_MAX_PACKET_LEN + 2]C.uchar
 	var flen = ax25_pack(pp, &fbuf[0])
-	return (ax25_only_hdlc_send_frame(channel, &fbuf[0], flen, bad_fcs))
+	return (ax25_only_hdlc_send_frame(channel, &fbuf[0], C.int(flen), bad_fcs))
 }
 
 func ax25_only_hdlc_send_frame(channel C.int, _fbuf *C.uchar, flen C.int, bad_fcs C.int) C.int {
